ref: Move KafkaConsumerWithCommitLog out of Kafka backend
This change separates the consumer with commit log from the base Kafka
consumer. The primary motivation is to eventually facilitate moving the
base consumer into a shared library (with Sentry). Since the consumer with
commit log is Snuba specific, it will stay in this codebase.
